requirements.txt
pip
PyOpenGL
GLFW
NumPy
Pyrr
Pillow
PyAssimp
Cython